Prevent unwanted WAL writes during binary upgrade
We add two GUC parameters:
* binary_upgrade_mode is another way to set IsBinaryUpgrade for binaries
  that do not have the -b/--binary-upgrade option (e.g. vacuumdb)
* binary_upgrade_allow_wal_writes is a new flag to indicate that WAL
  writes are allowed in binary upgrade mode

In XLogBeginInsert(), we now error out when in binary upgrade mode
unless binary_upgrade_allow_wal_writes parameter is on. Accordingly, the
error message is adjusted to remove mention of recovery which now does
not always apply to disallowed WAL writes. New debug messages inform
about the reason for WAL writes being disallowed or allowed.

In pg_upgrade, we now run vacuumdb in binary upgrade mode and allow WAL
writes (pg_upgrade fails with "PANIC: cannot insert new WAL entries"
otherwise); this is done through PGOPTIONS because 'vacuumdb --all' does not
accept a connection string (for passing "-o '-c ...'").

The next (unresolved) issue is that, when a cluster is started in binary
upgrade mode (with -b option), the checkpointer now fails because it
wants to write WAL.