Prevent unwanted WAL writes during binary upgrade
In XLogBeginInsert(), we now error out when in binary upgrade mode
unless WAL writes is explicitly allowed, which is controlled by the new
binary_upgrade_allow_wal_writes GUC parameter. Accordingly, the
error message is adjusted to remove mention of recovery which now does
not always apply to disallowed WAL writes. New debug messages inform
about the reason for WAL writes being disallowed or allowed.

Previously XLogInsertAllowed() was mostly concerned about in-progress
recovery. Now it also checks for binary upgrade mode. However, this
might not be enough since some code paths trying to insert WAL records
(e.g. the checkpointer) also special-case for in-progress recovery and
it might be needed to adjust them as well.

Now, when a cluster is started in binary upgrade mode (with -b option),
it now fails when starting child processes (e.g. the background writer
or the checkpointer?). To see this:

  $ ./src/backend/postgres -D pgdata -c backtrace_functions=XLogBeginInsert -b

(for details about backtrace_functions, see
https://amitdkhan-pg.blogspot.com/2020/07/backtraces-in-postgresql.html)

TODO:
* How to allow WAL writes from the checkpointer? Follow special-case
  handling for in-progress recovery?
* Is there a way to allow WAL writes from "internal" processes only
  (i.e. postmaster and maintenance workers)?