Grant/revoke role membership in a database
This adds a 'IN DATABASE <datname>' clause to GRANT/REVOKE commands
allowing to control role membership in a database scope. A major use
case is probably in combination with predefined roles, e.g.:

  GRANT pg_read_all_data TO bob IN DATABASE app;
  GRANT pg_maintain TO dba IN DATABASE metrics;

to avoid granting too many privileges when a user is supposed to only
operate on a subset of the cluster's databases.

The information about which database a role membership applies to is
stored in a new 'dbid' column of pg_auth_members catalog. When this
column is 0 (InvalidOid), the membership applies cluster-wise. Indexes
on this table are modified in order to include the new column to match
actual queries.

In roles_is_member_of(), the membership lookup is now performed first
for the database at stake, and then possibly repeated without a database
id for the cluster-wise case. The repetition of this operation is
factored out into a new append_role_memberships() helper function where
the previous logic now lives.

Function has_privs_of_role() now accepts a database id, as last
argument. For cases where a cluster-wise membership is sought, e.g. for
roles like 'pg_read_server_files', the new has_cluster_privs_of_role()
macro is used. The is_admin_of_role() function is modified similarly.

Original author: Kenaniah Cerny <kenaniah@gmail.com>.