Avoid using a superuser in tests
This is to avoid changing too much previous tests expectations as when
using a superuser, the grantor is in general resolved to the bootstrap
user (probably a "recent" change).

Consequently, when listing privileges (through check_memberships()), we
now get rows for roles created by 'role_admin', due to the fact that the
role creator is automatically granted ADMIN permission on the role.

To avoid displaying superuser name (which depends on environment), we
replace the value with a placeholder.

However, this poses problems in the following as we can no longer do
'SET ROLE'...